chore: moved types to types file
