+++ apps/multisig/src/components/TransactionSidesheet/TransactionSidesheetApprovals.tsx
@@ -1,7 +1,7 @@
 import MemberRow from '@components/MemberRow'
 import StatusCircle, { StatusCircleType } from '@components/StatusCircle'
 import { Tooltip } from '@components/ui/tooltip'
-import { Transaction } from '@domains/multisig'
+import { Transaction } from '@domains/offchain-data/metadata/types'
 import { useKnownAddresses } from '@hooks/useKnownAddresses'
 import { Address } from '@util/addresses'
 import { useSelectedMultisig } from '@domains/multisig'
