+++ apps/multisig/src/components/TransactionSidesheet/index.tsx
@@ -1,4 +1,5 @@
-import { Transaction, executingTransactionsState, useSelectedMultisig } from '@domains/multisig'
+import { executingTransactionsState, useSelectedMultisig } from '@domains/multisig'
+import { Transaction } from '@domains/offchain-data/metadata/types'
 import { Button, SideSheet } from '@talismn/ui'
 import { TransactionSidesheetHeader } from './TransactionSidesheetHeader'
 import { useCallback, useMemo, useState } from 'react'
@@ -15,7 +16,7 @@ import { useDeleteDraftMetadata, useSaveDraftMetadata } from '@domains/offchain-
 import { useNavigate } from 'react-router-dom'
 import { useToast } from '@components/ui/use-toast'
 import { useRecoilState, useSetRecoilState } from 'recoil'
-import { TxMetadata } from '@domains/offchain-data'
+import { TxMetadata } from '@domains/offchain-data/metadata/types'
 import { unknownConfirmedTransactionsState } from '@domains/tx-history'
 import { makeTransactionID } from '@util/misc'
 import { MIN_MULTISIG_THRESHOLD } from '@util/constants'
