+++ apps/multisig/src/layouts/NewTransaction/Vote/PendingVotes.tsx
@@ -9,7 +9,7 @@ import PendingVotesTable from './PendingVotesTable'
 import AmountRow from '@components/AmountRow'
 import { VotePill } from '../../../layouts/Overview/Transactions/VoteTransactionDetails'
 import { Multisig } from '@domains/multisig'
-import { Transaction } from '@domains/multisig'
+import { Transaction } from '@domains/offchain-data/metadata/types'
 import { ColumnDef } from '@tanstack/react-table'
 import BN from 'bn.js'
 import { SupportedChainIds } from '@domains/chains/generated-chains'
