+++ apps/multisig/src/layouts/Overview/Transactions/TransactionDetailsExpandable.tsx
@@ -8,7 +8,9 @@ import AmountRow from '@components/AmountRow'
 import MemberRow from '@components/MemberRow'
 import { decodeCallData } from '@domains/chains'
 import { pjsApiSelector, useApi } from '@domains/chains/pjs-api'
-import { Balance, Transaction, TransactionType, calcSumOutgoing, tempCalldataState } from '@domains/multisig'
+import { calcSumOutgoing, tempCalldataState } from '@domains/multisig'
+import { Transaction, Balance } from '@domains/offchain-data/metadata/types'
+import { TransactionType } from '@domains/offchain-data/metadata/types'
 import { Check, Contract, Copy, List, Send, Settings, Share2, Unknown, Users, Vote, Zap } from '@talismn/icons'
 import { useEffect, useMemo, useRef, useState } from 'react'
 import AceEditor from 'react-ace'
