+++ apps/multisig/src/layouts/SmartContracts/DeployContractExpandedDetails.tsx
@@ -1,4 +1,4 @@
-import { Transaction } from '@domains/multisig'
+import { Transaction } from '@domains/offchain-data/metadata/types'
 import { useMemo } from 'react'
 import { hexToU8a, compactAddLength, BN_ZERO } from '@polkadot/util'
 import { StatusMessage } from '@components/StatusMessage'
