+++ apps/multisig/src/layouts/Staking/ValidatorsRotationSummaryDetails.tsx
@@ -6,7 +6,8 @@ import AddressTooltip from '@components/AddressTooltip'
 import { Chain } from '@domains/chains'
 import { Validator, validatorsState } from '@domains/staking/ValidatorsWatcher'
 import { shortenAddress } from '@util/addresses'
-import { Transaction, useSelectedMultisig } from '@domains/multisig'
+import { useSelectedMultisig } from '@domains/multisig'
+import { Transaction } from '@domains/offchain-data/metadata/types'
 import { useNomPoolOf } from '@domains/staking/useNomPool'
 import { useNominations } from '@domains/staking/useNominations'
